Move Profiler overhead statistics rendering into a new component
