+++ src/components/app/MenuButtons/MetaOverheadStatistics.css
@@ -1,0 +1,33 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+.metaInfoGrid {
+  display: grid;
+  margin-bottom: 20px;
+  grid-template-columns: repeat(4, 1fr);
+}
+
+.metaInfoGrid > div {
+  margin: 2px 0;
+}
+
+.metaInfoGrid > div:nth-child(-n + 4) {
+  display: block;
+  font-weight: 500;
+}
+
+.metaInfoGrid > div:nth-child(4n + 1) {
+  color: var(--grey-50);
+  font-weight: 500;
+}
+
+.metaInfoGrid > div:not(:nth-child(4n + 1)) {
+  text-align: right;
+}
+
+.metaInfoValueRight {
+  display: inline-block;
+  width: calc(100% - 140px);
+  text-align: right;
+}
